`trait Pixels`: Move `len`, `as_ptr`-type methods from `Rav1dPictureDataComponent` to a `trait`.
